test(app): keep dev-smoke policy tests out of Playwright
